requirements.txt
Flask==2.2.3
Flask-Login==0.6.2
Werkzeug==2.2.3
pandas==1.5.3
nltk==3.8.1
numpy==1.24.2
gunicorn==20.1.0
python-dotenv==1.0.0
bcrypt==4.0.1
gensim==4.3.2
scikit-learn==1.3.0
scipy==1.11.3
Flask-Mail==0.9.1
python-jose==3.3.0
